Fix cyclic dependencies by moving some types to the domain layer
