+++ src/views/home/components/bridge-form/bridge-form.view.tsx
@@ -8,10 +8,10 @@ import { getEtherToken } from "src/constants";
 import { useEnvContext } from "src/contexts/env.context";
 import { useProvidersContext } from "src/contexts/providers.context";
 import { useTokensContext } from "src/contexts/tokens.context";
-import { Chain, FormData, Token } from "src/domain";
+import { AsyncTask, Chain, FormData, Token } from "src/domain";
 import useCallIfMounted from "src/hooks/use-call-if-mounted";
 import { isTokenEther, selectTokenAddress } from "src/utils/tokens";
-import { AsyncTask, isAsyncTaskDataAvailable } from "src/utils/types";
+import { isAsyncTaskDataAvailable } from "src/utils/types";
 import AmountInput from "src/views/home/components/amount-input/amount-input.view";
 import useBridgeFormStyles from "src/views/home/components/bridge-form/bridge-form.styles";
 import TokenSelector from "src/views/home/components/token-selector/token-selector.view";
