+++ src/views/home/components/token-list/token-list.view.tsx
@@ -6,10 +6,9 @@ import { ReactComponent as InfoIcon } from "src/assets/icons/info.svg";
 import { ReactComponent as MagnifyingGlassIcon } from "src/assets/icons/magnifying-glass.svg";
 import { ReactComponent as XMarkIcon } from "src/assets/icons/xmark.svg";
 import { useTokensContext } from "src/contexts/tokens.context";
-import { Chain, Token } from "src/domain";
+import { AsyncTask, Chain, Token } from "src/domain";
 import useCallIfMounted from "src/hooks/use-call-if-mounted";
 import { isTokenEther, selectTokenAddress } from "src/utils/tokens";
-import { AsyncTask } from "src/utils/types";
 import useTokenListStyles from "src/views/home/components/token-list/token-list.styles";
 import TokenSelectorHeader from "src/views/home/components/token-selector-header/token-selector-header.view";
 import Icon from "src/views/shared/icon/icon.view";
